Various bugfixes to spellbook screen.
Arcane and Divine tabs did nothing for a multiclass character.  They're
slightly functional now, with the possibly unwanted side effect of letting
spells per level count separately for arcane versus divine.  Multiclassing
still needs a lot of work; spellcaster code checks player level, not class
level, so a wizard 1 / cleric 1 has the spellcasting of a wizard 2 + cleric 2.

Disallow right-clicking to reduce the number of spells prepared below 0.

Tabs and spells per level were as wide as the dialog, which meant they
overlapped (and prevented the use of) the alphabetized spell list.

Charges display was too narrow, which meant it didn't display for some spells
(like a wizard's Summon Monster I).

Clicking on the alphabetized spell list caused an error, because fct was
unimplemented.  I implemented a simple fct to behave the same as clicking on a
spell's icon.  More cleanup would be good; it would be nice if the
alphabetical spell list and by-level spell list weren't in such different
formats, so code could be more easily reused.