Various fixes.
Reduce the width of spell image lists in the spellbook. As it was, they
overlapped the text spell list and prevented clicking on spells in that list.

The "Quit Permanently?" dialog can now only be closed by clicking Quit, not by
pressing Esc or clicking outside of the dialog.

Fix some misspellings. Try to improve the wording of one or two descriptions.
I'm leaving Resilient misspelled in its short_name because I think this will
help savegame compatibility; I'm not certain.

Canceling summoning a mount caused a Lua error.

Xavias's godPulse caused a Lua error.