Update Summon Mount to use proper dialog.
In T-Engine, registering a dialog returns immediately; it doesn't wait for the
dialog to finish.

A symptom of this is that Summon Mount starts the targeting interface
immediately, in the background of the mount selection dialog, instead of
waiting for the user to select a mount.

The standard way of handling this is by using coroutines. I still don't quite
get coroutines, so I took the talentDialog function from T-Engine 1.2.4 and
patched it back into this release and updated Summon Mount to use this.