Update Wild Shape to use talentDialog.
This fixes bugs such as an enemy getting a turn before the shape is selected or
the Wild Shape sustain starting even if you cancel a selection.

Update GetChoice to pass additional data to its callback. This lets a single
table (as in the table of wild shape options) contain all necessary data,
instead of duplicating code.